+++ src/frontend/src/eth/components/loaders/LoaderMultipleEthTransactions.svelte
@@ -2,12 +2,13 @@
 	import { debounce, isNullish } from '@dfinity/utils';
 	import { ETHERSCAN_MAX_CALLS_PER_SECOND } from '$env/rest/etherscan.env';
 	import { enabledEthereumTokens } from '$eth/derived/tokens.derived';
-	import { batchLoadEthereumTransactions } from '$eth/services/eth-transactions.services';
 	import { enabledErc20Tokens } from '$lib/derived/tokens.derived';
 	import type { TokenId } from '$lib/types/token';
+	import { batchLoadEthereumTransactions } from '$eth/services/eth-transactions-batch.services';
 
 	// TODO: make it more functional
 	let tokensLoaded: TokenId[] = [];
+
 
 	const load = async () => {
 		if (isNullish($enabledEthereumTokens) || isNullish($enabledErc20Tokens)) {
@@ -25,7 +26,7 @@
 		});
 
 		for await (const _ of loader) {
-			// The `yield` in the generator allows for asynchronous control flow
+			// This loop is just to wait for the loader to finish
 		}
 	};
